Refactor and update README
