+++ package.json
@@ -1,10 +1,11 @@
 {
-  "name": "react-route-generator",
-  "version": "0.0.27",
-  "description": "Route generator",
+  "name": "route-codegen",
+  "version": "0.0.1",
+  "description": "Route generator written in Typscript for React Router, NextJS, NodeJS and more",
   "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "bin": {
-    "routegen": "./dist/index.js"
+    "route-codegen": "./dist/bin/route-codegen.js"
   },
   "files": [
     "dist/**/*"
@@ -19,6 +20,7 @@
   "keywords": [
     "react-router",
     "next-js",
+    "route",
     "generator"
   ],
   "license": "MIT",
