+++ src/generate/utils/isNormalPattern/index.ts
