+++ src/generate/utils/isNormalPattern/isNormalPattern.ts
