+++ src/generateUrl/index.ts
@@ -1,0 +1,2 @@
+export { default } from './generateUrl';
+export * from './generateUrl';
