+++ src/index.ts
@@ -1,20 +1,3 @@
-#!/usr/bin/env node
-
-import yargs from 'yargs';
-import generate from './generate';
-import yaml from 'js-yaml';
-import { readFileSync } from 'fs';
-
-const argv = yargs.options({
-  config: { type: 'string', default: 'routegen.yml' },
-}).argv;
-
-try {
-  const { config } = argv;
-  const ymlContent = readFileSync(config, 'utf8');
-
-  const configContent = yaml.safeLoad(ymlContent);
-  generate(configContent);
-} catch (e) {
-  console.log(e);
-}
+export { default as generateUrl } from './generateUrl';
+export { default as generate } from './generate';
+export { Config, RoutingType } from './generate';
